# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ProducImage:
    properties:
      imageUrl:
        type: string
    required:
    - imageUrl
    type: object
  dto.ProductForm:
    properties:
      category:
        type: string
      comboProductsIds:
        items:
          type: integer
        type: array
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/dto.ProducImage'
        type: array
      name:
        type: string
      price:
        type: number
    required:
    - category
    - description
    - images
    - name
    - price
    type: object
  dto.ProductResponse:
    properties:
      category:
        type: string
      comboProducts:
        items:
          $ref: '#/definitions/dto.ProductResponse'
        type: array
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/dto.ProducImage'
        type: array
      name:
        type: string
      price:
        type: number
    required:
    - category
    - description
    - images
    - name
    - price
    type: object
...
host: localshot:3210
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Tech1 Fiap Project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tech1 API Docs
  version: "1.0"
paths:
...
  /api/admin/products:
    post:
      consumes:
      - application/json
      description: Create new product
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.ProductForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Product has required fields
        "409":
          description: This Product is already added
      summary: Create new product
      tags:
      - Product
  /api/admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update a product by ID
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Update a product
      tags:
      - Product
...
  /api/products/{id}:
    get:
      consumes:
      - application/json
      description: Get product by ID
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
      summary: Get product by ID
      tags:
      - Product
  /api/products/categories:
    get:
      consumes:
      - application/json
      description: Get all categories to filter in products by category
      parameters:
      - description: "12"
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Get all categories
      tags:
      - Product
  /api/products/categories/{category}:
    get:
      consumes:
      - application/json
      description: List all products by a category
      parameters:
      - description: Lanches
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductResponse'
            type: array
      summary: List all products by a category
      tags:
      - Product
...
swagger: "2.0"